requirements.txt
openai
requests
PyYAML
speechrecognition
pandas
vosk
pydub
ffmpeg-python
python-docx